# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CPCMS \u2014 Course Project Configuration and Management System",
    "description": "Internal University Course Project SCM & Configuration Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_v1_projects__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_api_v1_projects__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}/baselines": {
      "get": {
        "tags": [
          "Baselines"
        ],
        "summary": "List Baselines",
        "operationId": "list_baselines_api_v1_teams__team_id__baselines_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BaselineOut"
                  },
                  "title": "Response List Baselines Api V1 Teams  Team Id  Baselines Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Baselines"
        ],
        "summary": "Create Baseline",
        "operationId": "create_baseline_api_v1_teams__team_id__baselines_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/baselines/{id}/approve": {
      "post": {
        "tags": [
          "Baselines"
        ],
        "summary": "Approve Baseline",
        "operationId": "approve_baseline_api_v1_baselines__id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaselineCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "ci_version_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ci Version Ids"
          }
        },
        "type": "object",
        "required": [
          "name",
          "ci_version_ids"
        ],
        "title": "BaselineCreate"
      },
      "BaselineOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Name"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By Name"
          },
          "locked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CIVersionOut"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "code",
          "name",
          "status",
          "created_by",
          "created_at"
        ],
        "title": "BaselineOut"
      },
...
      "CITemplateOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ci_type": {
            "type": "string",
            "title": "Ci Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "position",
          "name",
          "ci_type",
          "created_at"
        ],
        "title": "CITemplateOut"
      },
...
      "DeadlineOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "kind",
          "title",
          "due_at",
          "created_at"
        ],
        "title": "DeadlineOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutcomeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "position",
          "text",
          "created_at"
        ],
        "title": "OutcomeOut"
      },
...
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "course_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "faculty_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faculty User Id"
          },
          "faculty_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "semester": {
            "type": "string",
            "title": "Semester"
          },
          "academic_year": {
            "type": "string",
            "title": "Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectives"
          },
          "min_team_size": {
            "type": "integer",
            "title": "Min Team Size"
          },
          "max_team_size": {
            "type": "integer",
            "title": "Max Team Size"
          },
          "team_formation_mode": {
            "type": "string",
            "title": "Team Formation Mode"
          },
          "team_formation_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Deadline"
          },
          "max_projects_per_student": {
            "type": "integer",
            "title": "Max Projects Per Student"
          },
          "allowed_file_exts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed File Exts"
          },
          "max_file_mb": {
            "type": "integer",
            "title": "Max File Mb"
          },
          "late_policy": {
            "type": "string",
            "title": "Late Policy"
          },
          "late_penalty_percent": {
            "type": "number",
            "title": "Late Penalty Percent"
          },
          "versioning_required": {
            "type": "boolean",
            "title": "Versioning Required"
          },
          "git_required": {
            "type": "boolean",
            "title": "Git Required"
          },
          "baseline_frequency": {
            "type": "string",
            "title": "Baseline Frequency"
          },
          "change_request_required": {
            "type": "boolean",
            "title": "Change Request Required"
          },
          "approval_required": {
            "type": "boolean",
            "title": "Approval Required"
          },
          "branching_policy": {
            "type": "string",
            "title": "Branching Policy"
          },
          "total_marks": {
            "type": "integer",
            "title": "Total Marks"
          },
          "grade_bands": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Grade Bands"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/ThemeOut"
            },
            "type": "array",
            "title": "Themes",
            "default": []
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeOut"
            },
            "type": "array",
            "title": "Outcomes",
            "default": []
          },
          "ci_templates": {
            "items": {
              "$ref": "#/components/schemas/CITemplateOut"
            },
            "type": "array",
            "title": "Ci Templates",
            "default": []
          },
          "deadlines": {
            "items": {
              "$ref": "#/components/schemas/DeadlineOut"
            },
            "type": "array",
            "title": "Deadlines",
            "default": []
          },
          "evaluation_criteria": {
            "items": {
              "$ref": "#/components/schemas/CriteriaOut"
            },
            "type": "array",
            "title": "Evaluation Criteria",
            "default": []
          },
          "user_team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Id"
          },
          "user_team_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Number"
          },
          "user_team_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "faculty_user_id",
          "name",
          "semester",
          "academic_year",
          "min_team_size",
          "max_team_size",
          "team_formation_mode",
          "max_projects_per_student",
          "allowed_file_exts",
          "max_file_mb",
          "late_policy",
          "late_penalty_percent",
          "versioning_required",
          "git_required",
          "baseline_frequency",
          "change_request_required",
          "approval_required",
          "branching_policy",
          "total_marks",
          "grade_bands",
          "status",
          "created_at"
        ],
        "title": "ProjectOut"
      },
...
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "semester": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "academic_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectives"
          },
          "min_team_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Team Size"
          },
          "max_team_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Team Size"
          },
          "team_formation_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Mode"
          },
          "team_formation_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Deadline"
          },
          "max_projects_per_student": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Projects Per Student"
          },
          "allowed_file_exts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed File Exts"
          },
          "max_file_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max File Mb"
          },
          "late_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Late Policy"
          },
          "late_penalty_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Late Penalty Percent"
          },
          "versioning_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versioning Required"
          },
          "git_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Required"
          },
          "baseline_frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Frequency"
          },
          "change_request_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Request Required"
          },
          "approval_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Required"
          },
          "branching_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branching Policy"
          },
          "total_marks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          },
          "grade_bands": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Bands"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "ProjectUpdate"
      },
...
      "ThemeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "created_at"
        ],
        "title": "ThemeOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}